feat: Refactor enemy imports and add Ghost and LandGhost classes
