+++ src/levels/Level.ts
@@ -1,7 +1,7 @@
-import { LandGhost } from "@/objects/LandGhost";
+import { Ghost } from "@/objects/enemies/Ghost";
+import { LandGhost } from "@/objects/enemies/LandGhost";
 import type { GameState } from "../engine/GameState";
 import { DiagonalPlatform } from "../objects/diagonalPlatform";
-import { Ghost } from "../objects/Ghost";
 import { MemoryCrystal } from "../objects/memoryCrystal";
 import { Platform } from "../objects/platform";
 import { Player } from "../objects/player";
