+++ src/levels/LevelEditor.ts
@@ -1,9 +1,9 @@
 import type { GameState } from "@/engine/GameState";
 import { Vector2 } from "@/engine/Vector2";
-import type { MemoryCrystal } from "@/objects/memoryCrystal";
 import type { DiagonalPlatform } from "@/objects/diagonalPlatform";
-import type { Ghost } from "@/objects/Ghost";
-import type { LandGhost } from "@/objects/LandGhost";
+import type { Ghost } from "@/objects/enemies/Ghost";
+import type { LandGhost } from "@/objects/enemies/LandGhost";
+import type { MemoryCrystal } from "@/objects/memoryCrystal";
 import type { Platform } from "@/objects/platform";
 import type { SolidBlock } from "@/objects/solidBlock";
 import { EditorLevelSaver } from "./LevelEditor/EditorLevelSaver";
@@ -71,7 +71,7 @@ export class LevelEditor {
       objectManager: this.objectManager,
       utils: this.utils,
     });
-    
+
     // Set initial crystal type
     this.mouseHandler.setCrystalType(this.crystalType);
