+++ src/objects/enemies/enemy.ts
@@ -1,5 +1,5 @@
-import { GameObject } from "../engine/GameObject";
-import type { GameState } from "../engine/GameState";
+import { GameObject } from "@/engine/GameObject";
+import type { GameState } from "@/engine/GameState";
 
 export abstract class Enemy extends GameObject {
   damage: number = 1;
