+++ src/systems/GameObjectManager.ts
@@ -1,7 +1,7 @@
+import type { Enemy } from "@/objects/enemies/enemy";
 import type { GameState } from "../engine/GameState";
 import type { ICollidable, IRenderable, IUpdateable } from "../interfaces/GameInterfaces";
 import type { DiagonalPlatform } from "../objects/diagonalPlatform";
-import type { Enemy } from "../objects/enemy";
 import type { Experience } from "../objects/experience";
 import { HitSpark, PoofEffect } from "../objects/hitSpark";
 import type { MemoryCrystal } from "../objects/memoryCrystal";
